server/pyproject.toml
[tool.poetry]
name = "Audapolis Server"
version = "0.0.1"
description = ""
authors = [
    "pajowu <audapolis@ca.pajowu.de>",
    "Jaro Habiger <jarohabiger@googlemail.com>",
]

[[tool.poetry.packages]]
from = "app"
include = "**/*.py"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.85.0"
uvicorn = { extras = ["standard"], version = "^0.18.0" }
python-multipart = "^0.0.5"

appdirs = "^1.4.4"
requests = "^2.26.0"
# OpenTimelineIO = "^0.14.0"  # Build fails on Apple Silicon, video export disabled
pydantic = "^1.8.2"
python_speech_features = "^0.6"
numpy = ">=1.22"
scipy = "^1.7.3"
webrtcvad = "^2.0.10"
scikit-learn = "^1.0.1"

setuptools = "<60.0.0"
openai-whisper = { git = "https://github.com/openai/whisper.git" }
torch = "^2.0.0"
torchaudio = "^2.0.0"

pyannote-audio = { version = "^3.1.0", extras = ["torch"] }
huggingface_hub = "^0.20.0"
loguru = "^0.7.3"



[tool.poetry.group.dev.dependencies]
isort = "^5.9.3"
black = "^21.8b0"
flake8 = "6.0.0"
rich = "^12.0.0"
types-requests = "^2.25.6"
types-PyYAML = "^5.4.10"
mypy = "^0.910"
beautifulsoup4 = "^4.10.0"
pycodestyle = "2.10.0"
pyflakes = "3.0.1"
mccabe = "0.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"
